refactor: оптимізація відображення пошуку та виправлення типу даних для населення

+++ GeographerDirectory/Models/City.cs
@@ -6,7 +6,7 @@
 
         public City() { }
 
-        public City(string name, int population, string coordinates)
+        public City(string name, long population, string coordinates)
             : base(name, population)
         {
             Coordinates = coordinates;
